pyproject.toml
[project]
name = "google-keep-mcp"
version = "1.0.0"
description = "MCP Server for Google Keep"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "gkeepapi>=0.16.0",
    "mcp[cli]",
    "python-dotenv",
]
authors = [
    { name = "Dave Nicoll", email = "d.nicoll@gmail.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 5 - Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/davenicoll/google-keep-mcp"
Repository = "https://github.com/davenicoll/google-keep-mcp"

[project.scripts]
mcp = "src.server.cli:main"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]